backend/requirements.txt
# Generated from @specs/features/authentication/spec.md
# FastAPI Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication - JWT verification
python-jose[cryptography]>=3.3.0
httpx>=0.27.0  # For async HTTP requests to JWKS endpoint

# Environment
python-dotenv>=1.0.0

# Validation and utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# CORS
starlette>=0.35.0
